Fixed Mobile View
the canvas size will adjust based on the screen size of the mobile device. For screens up to 600px wide, the canvas will take up the full width. For screens between 601px and 1024px, the canvas will take up 80% of the width. And for screens larger than 1024px, the canvas will have a fixed width of 800px and height of 600px.

This way, the canvas will be more suitable for mobile devices and adapt to different screen sizes.